refactor: 시큐리티 디렉토리 리펙토링
