+++ src/main/java/com/soongsil/CoffeeChat/security/dto/UserDto.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class UserDTO {
+public class UserDto {
     private String role;
     private String name;
     private String username; // 스프링 서버 내의 유저 아이디
